Refactor telemetry handlers and add no-op implementations
Removed TelemetryFactory.Handlers.swift and replaced individual metric handler classes with a unified CKTelemetryHandler supporting counter, floating point counter, and timer protocols. Introduced NoOpMeter and NoOpRecorder for meter and recorder metrics. Updated TelemetryFactory to use new handler classes for improved maintainability and clarity.
